% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIntensity.R
\name{get.intensity}
\alias{get.intensity}
\title{Get intensity}
\usage{
get.intensity(input, x, y, z, verbose = TRUE)
}
\arguments{
\item{input}{input a character vector consisting of the full path name to raw TIFF monochannel z-stack.}

\item{x}{pixel coordinates in x (0 is top left corner).}

\item{y}{pixel coordinates in y (0 is top left corner).}

\item{z}{pixel coordinates in z (0 is first plane)}

\item{verbose}{boolean value. If true diagnostic output is written to the R console. Deafult is true.}
}
\description{
Get the intensity from pixel coordinates of a z-stack
}
\examples{
#path to image
image<-'/Volumes/microscope/animal001/slide001/section001.tif'
#get the intensity
intensity <- get.intensity(image, x, y, z)
}
